refactor: move PhelFnLoaderInterface to Domain layer
