+++ tests/php/Unit/Internal/Domain/PhelFnNormalizerTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PhelTest\Unit\Internal\Domain;
 
 use Phel\Api\Application\PhelFnNormalizer;
-use Phel\Api\Infrastructure\PhelFnLoaderInterface;
+use Phel\Api\Domain\PhelFnLoaderInterface;
 use Phel\Api\Transfer\PhelFunction;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
 use PHPUnit\Framework\TestCase;
